backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 工具
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-slugify==8.0.1
markdown==3.5.2
feedgen==1.0.0
Pillow==10.2.0

# 开发
pytest==7.4.4
httpx==0.26.0
